requirements.txt
fastapi
uvicorn
catboost
pandas
joblib
sqlalchemy
python-dotenv